putting all the themes in one file
